feat: remove whitelist check in frontend and ask user to complete profile before creating agent
